+++ lib/src/core/escape/parser.dart
@@ -544,9 +544,11 @@ class EscapeParser {
           handler.setForegroundColor16(NamedColor.white);
           continue;
         case 38:
+          if (i + 1 >= params.length) break;
           final mode = params[i + 1];
           switch (mode) {
             case 2:
+              if (i + 4 >= params.length) break;
               final r = params[i + 2];
               final g = params[i + 3];
               final b = params[i + 4];
@@ -554,6 +556,7 @@ class EscapeParser {
               i += 4;
               break;
             case 5:
+              if (i + 2 >= params.length) break;
               final index = params[i + 2];
               handler.setForegroundColor256(index);
               i += 2;
@@ -589,9 +592,11 @@ class EscapeParser {
           handler.setBackgroundColor16(NamedColor.white);
           continue;
         case 48:
+          if (i + 1 >= params.length) break;
           final mode = params[i + 1];
           switch (mode) {
             case 2:
+              if (i + 4 >= params.length) break;
               final r = params[i + 2];
               final g = params[i + 3];
               final b = params[i + 4];
@@ -599,6 +604,7 @@ class EscapeParser {
               i += 4;
               break;
             case 5:
+              if (i + 2 >= params.length) break;
               final index = params[i + 2];
               handler.setBackgroundColor256(index);
               i += 2;
@@ -611,9 +617,11 @@ class EscapeParser {
 
         // CSI 58 - Underline color
         case 58:
+          if (i + 1 >= params.length) break;
           final mode = params[i + 1];
           switch (mode) {
             case 2:
+              if (i + 4 >= params.length) break;
               final r = params[i + 2];
               final g = params[i + 3];
               final b = params[i + 4];
@@ -621,6 +629,7 @@ class EscapeParser {
               i += 4;
               break;
             case 5:
+              if (i + 2 >= params.length) break;
               final index = params[i + 2];
               handler.setUnderlineColor256(index);
               i += 2;